requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
openai
requests
supabase
pandas
